+++ clients/linux/src/ui/mod.rs
@@ -1,6 +1,7 @@
 mod account_screen;
 mod onboard_screen;
 
+mod credit_card_input;
 mod filetree;
 mod menu_item;
 mod search_row;
@@ -20,6 +21,7 @@ pub use onboard_screen::OnboardOp;
 pub use onboard_screen::OnboardRoute;
 pub use onboard_screen::OnboardScreen;
 
+pub use credit_card_input::CreditCardInput;
 pub use filetree::FileTree;
 pub use filetree::FileTreeCol;
 pub use menu_item::menu_separator;
