+++ clients/linux/src/ui/usage_upgrade.rs
@@ -180,7 +180,7 @@ impl HeaderSection {
 struct SelectPayMethod {
     old_card: gtk::CheckButton,
     new_card: gtk::CheckButton,
-    new_card_input: CreditCardInput,
+    new_card_input: ui::CreditCardInput,
     btn_cancel: gtk::Button,
     btn_continue: gtk::Button,
     cntr: gtk::Box,
@@ -192,7 +192,7 @@ impl SelectPayMethod {
 
         let old_card = gtk::CheckButton::new();
 
-        let new_card_input = CreditCardInput::new();
+        let new_card_input = ui::CreditCardInput::new();
         let new_card = gtk::CheckButton::with_label("New Card");
 
         let methods = gtk::Box::new(gtk::Orientation::Vertical, 8);
@@ -305,120 +305,6 @@ impl ConfirmDetails {
             .margin_bottom(24)
             .build();
         self.content.append(&prompt);
-    }
-}
-
-struct CardInfo {
-    number: String,
-    cvc: String,
-    exp_month: i32,
-    exp_year: i32,
-}
-
-enum CardError {
-    Number,
-    Cvc,
-    ExpMonth,
-    ExpYear,
-}
-
-#[derive(Clone)]
-struct CreditCardInput {
-    number: gtk::Entry,
-    exp_month: gtk::Entry,
-    exp_year: gtk::Entry,
-    cvc: gtk::Entry,
-    error: gtk::Label,
-    revealer: gtk::Revealer,
-}
-
-impl CreditCardInput {
-    fn new() -> Self {
-        let error = gtk::Label::new(None);
-        error.add_css_class("err");
-        error.hide();
-
-        let card_input_entry = {
-            let error = error.clone();
-
-            move |placeholder: &str| {
-                let entry = gtk::Entry::new();
-                entry.set_placeholder_text(Some(placeholder));
-                entry.connect_changed({
-                    let error = error.clone();
-
-                    move |entry| {
-                        entry.remove_css_class("err-input");
-                        error.hide();
-                    }
-                });
-                entry
-            }
-        };
-
-        let number = card_input_entry("Card Number");
-        let exp_month = card_input_entry("MM");
-        let exp_year = card_input_entry("YY");
-        let cvc = card_input_entry("CVC");
-
-        number.set_width_request(260);
-
-        let inputs = gtk::Box::new(gtk::Orientation::Horizontal, 4);
-        inputs.append(&number);
-        inputs.append(&exp_month);
-        inputs.append(&exp_year);
-        inputs.append(&cvc);
-
-        let cntr = gtk::Box::new(gtk::Orientation::Vertical, 8);
-        cntr.append(&inputs);
-        cntr.append(&error);
-
-        let revealer = gtk::Revealer::builder().child(&cntr).build();
-
-        Self { number, cvc, exp_month, exp_year, error, revealer }
-    }
-
-    fn info(&self) -> Result<CardInfo, CardError> {
-        let number = self.number.text().to_string();
-        if number.len() < 14 || number.len() > 16 {
-            return Err(CardError::Number);
-        }
-
-        let cvc = self.cvc.text().to_string();
-        if cvc.len() < 3 {
-            return Err(CardError::Cvc);
-        }
-
-        let exp_month: i32 = self
-            .exp_month
-            .text()
-            .to_string()
-            .parse()
-            .map_err(|_| CardError::ExpMonth)?;
-        if !(1..=12).contains(&exp_month) {
-            return Err(CardError::ExpMonth);
-        }
-
-        let exp_year: i32 = self
-            .exp_year
-            .text()
-            .to_string()
-            .parse()
-            .map_err(|_| CardError::ExpYear)?;
-
-        Ok(CardInfo { number, cvc, exp_month, exp_year })
-    }
-
-    fn handle_err(&self, err: CardError) {
-        let (entry, msg) = match err {
-            CardError::Number => (&self.number, "Please enter a valid card number."),
-            CardError::Cvc => (&self.cvc, "Please enter a valid CVC."),
-            CardError::ExpMonth => (&self.exp_month, "Please enter a valid expiry month."),
-            CardError::ExpYear => (&self.exp_year, "Please enter a valid expiry year."),
-        };
-        entry.add_css_class("err-input");
-        self.error.set_text(msg);
-        self.error.show();
     }
 }
